fix: fix caps and change lang by mouse
